stake_updater/requirements.txt
python-dotenv==1.0.1
psycopg2-binary==2.9.9
substrate-interface==1.7.10
scalecodec==1.2.11
bt-decode
